+++ utils.js
@@ -1,0 +1,2 @@
+//Randomized HP function
+export const getRandom = (num) => Math.ceil(Math.random() * num);
